Fix crash when restoring state on TextFragment
`onViewStateRestored()` in TextFragment crashed with `ClassCastException`,
because layout has `MyEditText` with id `text_note_view` and not `MyTextView`.
This has now been changed to `TextView` to ensure less breaking in the future
and since only `text` was changed, this is enough.

This happened when viewing notes and causing configuration change or something else
that would cause the state to be restored.